requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
boto3==1.28.64
python-dotenv==1.0.0
requests==2.31.0
jsonschema==4.19.1
pydantic==2.4.2
python-multipart==0.0.6 